Fix the legacy error issue?
Moving the plugin implementation out of the test class and
removing the legacy injector make the test pass in maven.
